+++ Sources/Swift/Integrations/SessionReplay/SentryReplayVideoMaker.swift
@@ -7,8 +7,15 @@ protocol SentryReplayVideoMaker: NSObjectProtocol {
     var videoWidth: Int { get set }
     var videoHeight: Int { get set }
     
-    func addFrameAsync(image: UIImage)
+    func addFrameAsync(image: UIImage, forScreen: String?) 
     func releaseFramesUntil(_ date: Date)
-    func createVideoWith(duration: TimeInterval, beginning: Date, outputFileURL: URL, completion: @escaping (SentryVideoInfo?, Error?) -> Void) throws
+    func createVideoWith(beginning: Date, end: Date, outputFileURL: URL, completion: @escaping (SentryVideoInfo?, Error?) -> Void) throws
 }
+
+extension SentryReplayVideoMaker {
+    func addFrameAsync(image: UIImage) {
+        self.addFrameAsync(image: image, forScreen: nil)
+    }
+}
+
 #endif